Add ability to select and click on cat from list
Now there is a list of cats to choose from. When a
user clicks on one of the cats, that cat is put inside
the top div where the user can click on it to track
mouse clicks. The app then updates the number of clicks
in the headers both in the list and in the selected cat
cat.